requirements.txt
asgiref==3.8.1
Django==5.1.6
Faker==37.0.0
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.1
django-environs==0.1.1
djangorestframework==3.16.0
django-filter==25.1
drf-yasg==1.21.10
whitenoise==6.9.0
django-debug-toolbar==6.0.0
drf-nested-routers==0.94.2